+++ sleepecg/io/__init__.py
@@ -2,11 +2,13 @@
 
 from .ecg_readers import read_gudb, read_ltdb, read_mitdb
 from .nsrr import download_nsrr, set_nsrr_token
+from .sleep_readers import read_mesa
 
 __all__ = [
     'read_gudb',
     'read_ltdb',
     'read_mitdb',
+    'read_mesa',
     'set_nsrr_token',
     'download_nsrr',
 ]